check value of summands

+++ abramov.vladislav/P2/main.cpp
@@ -7,7 +7,7 @@ int main()
 {
   double left = 0.0;
   double right = 0.0;
-  size_t k = 0;
+  int k = 0;
   std::cin >> left >> right >> k;
   if (!std::cin)
   {
